fix python cmdline parser for variants
Given
```
type operation = Copy of string | Mirror of string
```

Represented as `["Copy", "arg"]` we need to parse that as JSON when given on the cmdline.

Otherwise it'd do `operation[0]`, where `operation` is a string, which is not what we want.

This doesn't affect usage from xapi-storage-script (which always calls it with --json),
but it affects testing the python code by hand, because you can't supply the correct cmdline arguments
(unless you construct an entire JSON by hand and pass it on stdin which is a bit tedious)

Signed-off-by: Edwin Török <edwin.torok@cloud.com>